[rs-dataprovider] first phase optim
Former-commit-id: a1149de362f98850f651376fad17a01ee0d7b981

+++ rs-dataprovider/acquisition/acquisition-plugins-ssalto/src/test/java/fr/cnes/regards/modules/acquisition/plugins/ssalto/chain/Jason2IgdrProcessingChainTest.java
@@ -127,12 +127,12 @@ public class Jason2IgdrProcessingChainTest extends AbstractAcquisitionChainTest 
     }
 
     @Override
-    protected int getExpectedFiles() {
+    protected long getExpectedFiles() {
         return 3;
     }
 
     @Override
-    protected int getExpectedProducts() {
+    protected long getExpectedProducts() {
         return 3;
     }
